find all

+++ src/main/java/FCI/graduate/blood_Donation/dto/DonorInfoDto.java
@@ -27,15 +27,15 @@ public class DonorInfoDto {
 
 	private String gender;
 
-	private int age;
+	private Integer age;
 
 	private String phone;
 
-	private int longitude;
+	private Double longitude;
 
-	private int latitude;
+	private Double latitude;
 
-	private int countDonates;
+	private Integer countDonates;
 
 	private String bloodType;
